client.lua: Don't send back messages
When the client receives a messages it will no longer send it back to
the server after applying the change. This fixes an issue that happened
when one person hit seek two or more times and the other clients
received the first seek, sent it back to the server after applying it
and a ping pong effect was created.